+++ CHANGELOG.md
@@ -13,6 +13,7 @@ speaking, considered to be the player's experience.
 ### Added
 
 - Added scroll up and scroll down prompts to the What's New screen.
+- Added scroll hints to the What's New screen.
 
 
 ## [3.1.9] — 2025-06-25